refactor(suite-common): extract bip329 label package
