+++ suite-common/bip329/src/index.ts
@@ -1,0 +1,1 @@
+export type { Bip329Label } from './bip329';
