+++ suite-common/metadata-types/src/metadataTypes.ts
@@ -294,10 +294,3 @@ export type PasswordManagerState = {
     extVersion?: string;
     tags: Record<number, PasswordTag>;
 };
-
-export type Bip329Label = {
-    type: 'tx' | 'addr' | 'wallet' | 'xpub' | 'pubkey' | 'input' | 'output';
-    ref?: string; // The identifier for the object being labeled (e.g., txid, address, txid:vout)
-    label: string; // The label text
-    spendable?: boolean;
-};
