+++ suite-common/suite-sync/package.json
@@ -13,11 +13,11 @@
     },
     "dependencies": {
         "@reduxjs/toolkit": "2.11.2",
+        "@suite-common/bip329": "workspace:*",
         "@suite-common/delegated-identity-key": "workspace:*",
         "@suite-common/delegated-identity-key-types": "workspace:*",
         "@suite-common/dependency-injection": "workspace:*",
         "@suite-common/device": "workspace:*",
-        "@suite-common/metadata-types": "workspace:*",
         "@suite-common/platform-encryption": "workspace:*",
         "@suite-common/redux-utils": "workspace:*",
         "@suite-common/suite-sync-quota-manager": "workspace:*",
