+++ suite-common/suite-sync/tsconfig.json
@@ -5,13 +5,13 @@
         "exactOptionalPropertyTypes": true
     },
     "references": [
+        { "path": "../bip329" },
         { "path": "../delegated-identity-key" },
         {
             "path": "../delegated-identity-key-types"
         },
         { "path": "../dependency-injection" },
         { "path": "../device" },
-        { "path": "../metadata-types" },
         { "path": "../platform-encryption" },
         { "path": "../redux-utils" },
         { "path": "../suite-sync-quota-manager" },
