+++ suite/metadata/src/slip15ToBip329.ts
@@ -1,4 +1,5 @@
-import { type AccountLabels, type Bip329Label } from '@suite-common/metadata-types';
+import { type Bip329Label } from '@suite-common/bip329';
+import { type AccountLabels } from '@suite-common/metadata-types';
 
 // Transforms a custom SLIP-15 like wallet label object into an array of BIP-329 label objects.
 export const slip15ToBip329 = (inputData: AccountLabels, allSpendable = true): Bip329Label[] => {
