+++ suite/metadata/tsconfig.json
@@ -11,6 +11,7 @@
     },
     "include": [".", "**/*.json"],
     "references": [
+        { "path": "../../suite-common/bip329" },
         { "path": "../../suite-common/device" },
         {
             "path": "../../suite-common/metadata-types"
